requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
textstat>=0.7.3
spacy>=3.7.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0

# Text Analysis
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.30.0

# Vector Operations
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Web Framework
requests>=2.31.0
urllib3>=2.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Azure Integration
azure-ai-textanalytics>=5.3.0
azure-storage-blob>=12.17.0

# Cohere Integration
cohere>=4.21.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0